업데이트수정

+++ src/main/java/com/project/member/MemberService.java
@@ -49,21 +49,25 @@ public class MemberService {
 	//회원 정보 수정
 	public int setUpdate(MemberVO memberVO, MultipartFile photo, HttpSession session) throws Exception{
 		MemberFileVO fileVO = memberFileDAO.getSelect(memberVO);
-		String realpath = session.getServletContext().getRealPath("/resources/member");
-		//System.out.println(realpath);
-		String fname = fileSaver.saveFile(realpath, photo);
-		String oname = photo.getOriginalFilename();
-		MemberFileVO memberFileVO = new MemberFileVO();
-		memberFileVO.setId(memberVO.getId());
-		memberFileVO.setFname(fname);
-		memberFileVO.setOname(oname);
-		
 		int result = memberDAO.setUpdate(memberVO);
-		result = memberFileDAO.setWrite(memberFileVO);
-		result = memberFileDAO.setDelete(fileVO);
-		if(result>0) {
-			fileSaver.deleteFile(realpath, fileVO.getFname());
+		
+		if(photo.getOriginalFilename().length()>0) {
+			String realpath = session.getServletContext().getRealPath("/resources/member");
+			//System.out.println(realpath);
+			String fname = fileSaver.saveFile(realpath, photo);
+			String oname = photo.getOriginalFilename();
+			MemberFileVO memberFileVO = new MemberFileVO();
+			memberFileVO.setId(memberVO.getId());
+			memberFileVO.setFname(fname);
+			memberFileVO.setOname(oname);
+			result = memberFileDAO.setWrite(memberFileVO);
+			result = memberFileDAO.setDelete(fileVO);
+			
+			if(result>0) {
+				fileSaver.deleteFile(realpath, fileVO.getFname());
+			}
 		}
+		
 		
 		return result;
 	}
